Prevent widgets being added to other widgets
tkinter allows any widget to be the parent container of another, meaning that a button can hold a text input for example. This seems like nonsensical behaviour, so all widgets now check they are either a ttk.Frame or ttk.LabelFrame which is the top of the inheritance tree for containers.