Change to Listbox widget and add listbox test app
Several changes made to listbox widget, including handling of getting values that are dependent on the mode of the listbox and adding several methods for selecting and altering the listbox items. Created the listbox testing app to ensure all calls to the listbox work as intended.

Also random small change to container 'add' function to fix an error involving the sticky property.