Change set default icon method
A report (unable to be reproduced) of the default window icon (GooeyPie logo) not working on Mac. The base64 encoded data was unable to be recognized as image data in the call to tkinter.PhotoImage.

The new method encodes a 128px PNG image and uses Pillow instead of tkinter to produce the icon. The icon remains at full size for Mac, since the icon shows in the dock, but is resized down to 32px for Windows and Linux.