Remove check for correct number of arguments in event function
To allow GooeyPie to function correctly in both an imperative and object-oriented approach, the check of whether a specified event function accepts a single argument has been removed, since an OOP method always receives a reference to the current object (self) in addition to any other arguments. During the execution of the app, if a function or method does not accept the correct number of arguments, a TypeError is caught and re-raised with an appropriate message.